Refactor route modules and page data loading
