+++ crates/web-server/handlers/api_keys/loader.rs
@@ -2,7 +2,7 @@ use crate::{CustomError, Jwt};
 use axum::extract::Extension;
 use axum::response::Html;
 use db::authz;
-use db::{queries, Pool};
+use db::Pool;
 use web_pages::{api_keys, routes::api_keys::Index};
 
 pub async fn loader(
@@ -20,40 +20,16 @@ pub async fn loader(
         return Err(CustomError::Authorization);
     }
 
-    let api_keys = queries::api_keys::api_keys()
-        .bind(&transaction, &team_id_num)
-        .all()
-        .await?;
-
-    let assistants = queries::prompts::prompts()
-        .bind(&transaction, &team_id_num, &db::PromptType::Assistant)
-        .all()
-        .await?;
-
-    let models = queries::prompts::prompts()
-        .bind(&transaction, &team_id_num, &db::PromptType::Model)
-        .all()
-        .await?;
-
-    // Fetch graph data for the last 7 days
-    let token_usage_data = queries::token_usage_metrics::get_daily_token_usage_for_team()
-        .bind(&transaction, &team_id_num, &"7")
-        .all()
-        .await?;
-
-    let api_request_data = queries::token_usage_metrics::get_daily_api_request_count_for_team()
-        .bind(&transaction, &team_id_num, &"7")
-        .all()
-        .await?;
+    let page_data = super::page_data::load_api_keys_page_data(&transaction, team_id_num).await?;
 
     let html = api_keys::page::page(
         rbac,
         team_id,
-        api_keys,
-        assistants,
-        models,
-        token_usage_data,
-        api_request_data,
+        page_data.api_keys,
+        page_data.assistants,
+        page_data.models,
+        page_data.token_usage_data,
+        page_data.api_request_data,
         None,
     );
