+++ crates/web-server/handlers/providers/actions.rs
@@ -1,137 +1,13 @@
 use crate::layout::{empty_string_is_none, empty_string_is_none_i32};
 use crate::{CustomError, Jwt};
 use axum::extract::Extension;
-use axum::response::Html;
 use axum::response::IntoResponse;
 use axum::Form;
-use axum::Router;
-use axum_extra::routing::RouterExt;
 use db::authz;
-use db::queries;
-use db::Pool;
+use db::{queries, Pool};
 use serde::{Deserialize, Deserializer};
 use validator::Validate;
-use web_pages::providers::upsert as provider_page;
-use web_pages::routes::providers::{Delete, Edit, Index, New, Upsert};
-
-pub fn routes() -> Router {
-    Router::new()
-        .typed_get(loader)
-        .typed_get(new_loader)
-        .typed_get(edit_loader)
-        .typed_post(upsert_action)
-        .typed_post(delete_action)
-}
-
-pub async fn loader(
-    Index { team_id }: Index,
-    current_user: Jwt,
-    Extension(pool): Extension<Pool>,
-) -> Result<Html<String>, CustomError> {
-    let mut client = pool.get().await?;
-    let transaction = client.transaction().await?;
-
-    let (rbac, _team_id_num) =
-        authz::get_permisisons(&transaction, &current_user.into(), &team_id).await?;
-
-    if !rbac.is_sys_admin {
-        return Err(CustomError::Authorization);
-    }
-
-    let providers = queries::providers::providers()
-        .bind(&transaction)
-        .all()
-        .await?;
-
-    let html = web_pages::providers::page::page(team_id, rbac, providers);
-
-    Ok(Html(html))
-}
-
-pub async fn new_loader(
-    New { team_id }: New,
-    current_user: Jwt,
-    Extension(pool): Extension<Pool>,
-) -> Result<Html<String>, CustomError> {
-    let mut client = pool.get().await?;
-    let transaction = client.transaction().await?;
-
-    let (rbac, _team_id_num) =
-        authz::get_permisisons(&transaction, &current_user.into(), &team_id).await?;
-
-    if !rbac.is_sys_admin {
-        return Err(CustomError::Authorization);
-    }
-
-    let form = provider_page::ProviderForm {
-        id: None,
-        name: "".to_string(),
-        svg_logo: "".to_string(),
-        default_model_name: "".to_string(),
-        default_model_display_name: "".to_string(),
-        default_model_context_size: 0,
-        default_model_description: "".to_string(),
-        base_url: "".to_string(),
-        api_key_optional: false,
-        default_embeddings_model_name: "".to_string(),
-        default_embeddings_model_display_name: "".to_string(),
-        default_embeddings_model_context_size: 0,
-        default_embeddings_model_description: "".to_string(),
-        error: None,
-    };
-
-    let html = provider_page::page(team_id, rbac, form);
-
-    Ok(Html(html))
-}
-
-pub async fn edit_loader(
-    Edit { team_id, id }: Edit,
-    current_user: Jwt,
-    Extension(pool): Extension<Pool>,
-) -> Result<Html<String>, CustomError> {
-    let mut client = pool.get().await?;
-    let transaction = client.transaction().await?;
-
-    let (rbac, _team_id_num) =
-        authz::get_permisisons(&transaction, &current_user.into(), &team_id).await?;
-
-    if !rbac.is_sys_admin {
-        return Err(CustomError::Authorization);
-    }
-
-    let provider = queries::providers::provider()
-        .bind(&transaction, &id)
-        .one()
-        .await?;
-
-    let form = provider_page::ProviderForm {
-        id: Some(provider.id),
-        name: provider.name,
-        svg_logo: provider.svg_logo,
-        default_model_name: provider.default_model_name.unwrap_or_default(),
-        default_model_display_name: provider.default_model_display_name.unwrap_or_default(),
-        default_model_context_size: provider.default_model_context_size,
-        default_model_description: provider.default_model_description,
-        base_url: provider.base_url,
-        api_key_optional: provider.api_key_optional,
-        default_embeddings_model_name: provider.default_embeddings_model_name.unwrap_or_default(),
-        default_embeddings_model_display_name: provider
-            .default_embeddings_model_display_name
-            .unwrap_or_default(),
-        default_embeddings_model_context_size: provider
-            .default_embeddings_model_context_size
-            .unwrap_or_default(),
-        default_embeddings_model_description: provider
-            .default_embeddings_model_description
-            .unwrap_or_default(),
-        error: None,
-    };
-
-    let html = provider_page::page(team_id, rbac, form);
-
-    Ok(Html(html))
-}
+use web_pages::routes::providers::{Delete, Upsert};
 
 #[derive(Deserialize, Validate, Default, Debug)]
 pub struct ProviderForm {
@@ -172,7 +48,7 @@ where
     ))
 }
 
-pub async fn upsert_action(
+pub async fn action_upsert(
     Upsert { team_id }: Upsert,
     current_user: Jwt,
     Extension(pool): Extension<Pool>,
@@ -252,7 +128,7 @@ pub async fn upsert_action(
     }
 }
 
-pub async fn delete_action(
+pub async fn action_delete(
     Delete { id, team_id }: Delete,
     current_user: Jwt,
     Extension(pool): Extension<Pool>,
