+++ crates/web-server/handlers/rate_limits/mod.rs
@@ -1,6 +1,5 @@
 mod actions;
 mod loader;
-mod page_data;
 
 pub use actions::*;
 pub use loader::*;
@@ -11,6 +10,6 @@ use axum_extra::routing::RouterExt;
 pub fn routes() -> Router {
     Router::new()
         .typed_get(loader::loader)
-        .typed_post(actions::action_new_api_key)
-        .typed_post(actions::action_delete_api_key)
+        .typed_post(actions::action_upsert)
+        .typed_post(actions::action_delete)
 }
